SideBar added pages, components, cleaned app.js, added files for different screens
