+++ src/frontend/covid-near-me/src/pages/CovidTestingCentre.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function CovidTestingCentres() {
+    return (
+        <div className='covid-testing-centre'>
+            Covid Testing Centre 
+        </div>
+    )
+}
+
+export default CovidTestingCentres
