+++ src/frontend/covid-near-me/src/pages/Vaccinations.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Vaccinations() {
+    return (
+        <div className='vaccination'>
+            Vaccination
+        </div>
+    )
+}
+
+export default Vaccinations
